Move send acceptance behind the Hono control plane
Extract the existing transactional send and batch orchestration into shared request handlers so services/api can expose the control-plane routes while Next.js remains a compatibility adapter.\n\nConstraint: Public OpenSend/Resend send API shape, idempotency conflicts, auth errors, validation envelopes, quota reservation, and queue-first behavior must remain compatible.\nRejected: Proxying Hono back to Next.js | would leave route ownership in the old runtime and fail the #279 service ownership goal.\nRejected: Full packages/core extraction in this slice | would force app-only seams across package boundaries before they are individually injectable.\nConfidence: high\nScope-risk: moderate\nDirective: Move send dependencies into core one seam at a time; keep Next and Hono adapters sharing the same handler until production routing cuts over.\nTested: make check; bun run test -- tests/api-emails.test.ts tests/api-service-health.test.ts; bun run build:api\nNot-tested: Full make test and Playwright e2e.
